+++ web-admin/src/features/projects/ProjectDashboardsListener.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { listenAndInvalidateDashboards } from "@rilldata/web-admin/features/dashboards/listing/dashboards";
+  import { listenAndInvalidateDashboards } from "@rilldata/web-admin/features/dashboards/listing/selectors";
   import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
   import { useQueryClient } from "@tanstack/svelte-query";
   import { onDestroy } from "svelte";
